refactor: NavButtons 컴포넌트 삭제
- 굳이 나눌 필요 없는 컴포넌트라고 판단하여 삭제
