--- FE/src/components/IssueList/IssueListHeader/NavButtons/style.js
@@ -1,12 +1,0 @@
-import styled from 'styled-components';
-
-const $NavButtons = styled.div`
-  display: flex;
-  align-items: center;
-
-  & > div:first-child {
-    margin-right: 16px;
-  }
-`;
-
-export { $NavButtons };
