+++ FE/src/components/IssueList/IssueListHeader/index.jsx
@@ -1,15 +1,28 @@
 import React from 'react';
+import { useNavigate } from 'react-router-dom';
 import PropTypes from 'prop-types';
 
-import NavButtons from './NavButtons';
 import FilterBar from './FilterBar';
-import { $IssueListHeader } from './style';
+import Button from '../../common/Button';
+import TabButton from '../../common/TabButton';
+import Icon from '../../common/Icon';
+import { $IssueListHeader, $Nav } from './style';
 
 const IssueListHeader = ({ labelCount, milestoneCount }) => {
+  const navigate = useNavigate();
+
+  const navigateToNewIssue = () => navigate('/issues/new');
+
   return (
     <$IssueListHeader>
       <FilterBar />
-      <NavButtons labelCount={labelCount} milestoneCount={milestoneCount} />
+      <$Nav>
+        <TabButton labelCount={labelCount} milestoneCount={milestoneCount} />
+        <Button type="contained" size="S" onClick={navigateToNewIssue}>
+          <Icon name="plus" />
+          이슈 작성
+        </Button>
+      </$Nav>
     </$IssueListHeader>
   );
 };
