+++ FE/src/components/IssueList/IssueListHeader/style.js
@@ -1,9 +1,15 @@
 import styled from 'styled-components';
 
+const $Nav = styled.nav`
+  display: flex;
+  align-items: center;
+  gap: 16px;
+`;
+
 const $IssueListHeader = styled.header`
   display: flex;
   justify-content: space-between;
   align-items: center;
 `;
 
-export { $IssueListHeader };
+export { $IssueListHeader, $Nav };
